+++ diive/pkgs/createvar/conversions.py
@@ -26,6 +26,9 @@ def air_temp_from_sonic_temp(sonic_temp: pd.Series, h2o: pd.Series) -> pd.Series
 
     Returns:
         pd.Series: Air temperature data in Kelvin.
+
+    Example:
+        See `examples/createvar/conversions.py` for complete examples.
     """
     ta = sonic_temp / (1 + 0.32 * h2o)
     ta.name = "TA_SONIC"
@@ -48,6 +51,9 @@ def latent_heat_of_vaporization(ta: pd.Series) -> pd.Series:
 
     Returns:
         Latent heat of vaporization (J kg-1)
+
+    Example:
+        See `examples/createvar/conversions.py` for complete examples.
     """
     k1 = 2.501
     k2 = 0.00237
@@ -68,43 +74,11 @@ def et_from_le(le: pd.Series, ta: pd.Series) -> pd.Series:
 
     Returns:
         Evapotranspiration (mm H2O h-1)
+
+    Example:
+        See `examples/createvar/conversions.py` for complete examples.
     """
     _lambda = latent_heat_of_vaporization(ta)
     et = le / _lambda  # kg m-2 s-1 = mm s-1
     et = et.multiply(3600)  # mm h-1
     return et
-
-
-def _example_et_from_le():
-    from diive.configs.exampledata import load_exampledata_EDDYPRO_FLUXNET_CSV_30MIN
-    df, meta = load_exampledata_EDDYPRO_FLUXNET_CSV_30MIN()
-    le = df['LE'].copy()
-    et_check = df['ET'].copy()  # Should be in mm h-1
-    ta = df['TA_1_1_1'].copy()
-
-    et = et_from_le(le=le, ta=ta)
-    print(et)
-    print(et_check)
-
-    import matplotlib.pyplot as plt
-    et.plot(label="ET")
-    et_check.plot(label="ET from EddyPro")
-    plt.legend()
-    plt.show()
-
-    # # Testing polars vs pandas
-    # # Execution time for 100 000 repeats:
-    # #     pandas: 11.505 seconds
-    # #     polars:  1.975 seconds
-    # ta = pl.from_pandas(ta)
-    # print(type(ta))
-    # start_time = time.perf_counter()
-    # for i in range(100_000):
-    #     res = latent_heat_of_vaporization(ta=ta)
-    # end_time = time.perf_counter()
-    # elapsed_time = end_time - start_time
-    # print(f"Elapsed time: {elapsed_time} seconds")
-
-
-if __name__ == '__main__':
-    _example_et_from_le()